backend/requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
nltk==3.8.1
scikit-learn==1.3.2
python-dotenv==1.0.0